refactor: move login to user routes
