+++ api/authentication/authRoute.js
@@ -1,8 +1,4 @@
 import express from 'express';
-import bcryptjs from 'bcryptjs';
-import jsonwebtoken from 'jsonwebtoken';
-import config from 'config';
-import { check, validationResult } from 'express-validator';
 import auth from './authMiddleware.js';
 import User from '../user/User.js';
 
@@ -20,69 +16,5 @@ router.get('/', auth, async (req, res) => {
     res.status(500).send('Server Error');
   }
 });
-
-// @route    POST api/auth
-// @desc     Authenticate user & get token
-// @access   Public
-router.post(
-  '/',
-  check('email', 'Please include a valid email').isEmail(),
-  check('password', 'Password is required').exists(),
-  async (req, res) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({ errors: errors.array() });
-    }
-
-    const { email, password } = req.body;
-
-    try {
-      const user = await User.findOne({ email });
-
-      if (!user) {
-        return res
-          .status(400)
-          .json({ errors: [{ msg: 'Invalid Credentials' }] });
-      }
-
-      const isMatch = await bcryptjs.compare(password, user.password);
-
-      if (!isMatch) {
-        return res
-          .status(400)
-          .json({ errors: [{ msg: 'Invalid Credentials' }] });
-      }
-
-      const payload = {
-        user: {
-          id: user.id,
-          isAdmin: user.isAdmin,
-        },
-      };
-
-      jsonwebtoken.sign(
-        payload,
-        config.get('jwtSecret'),
-        { expiresIn: '36000' },
-        (err, token) => {
-          if (err) throw err;
-          res
-            .cookie('access_token', token, {
-              httpOnly: true,
-              sameSite: true,
-            })
-            .json({
-              isAuthenticated: true,
-              user: { id: user.id, isAdmin: user.isAdmin },
-            })
-            .status(200);
-        },
-      );
-    } catch (err) {
-      console.error(err.message);
-      res.status(500).send('Server error');
-    }
-  },
-);
 
 export default router;
